Hand-roll image srcset
Might be taking a bit of a gamble on this one, since we're not being
nearly as careful as the Retcon plugin is, but Retcon alone can't handle
remote assets, so we're doing it this way.

A lot of this code is specific to this codebase, but it might be worth
trying to abstract it away in future.